usm: tests: classification: Move mysql classification test to USM linux file
